YQL-19747 Extract NormalizeName from yql/.../core/ast
