+++ yql/essentials/ast/yql_expr.h
@@ -9,6 +9,7 @@
 #include "yql_pos_handle.h"
 
 #include <yql/essentials/core/url_lister/interface/url_lister_manager.h>
+#include <yql/essentials/core/sql_types/normalize_name.h>
 #include <yql/essentials/utils/yql_panic.h>
 #include <yql/essentials/public/issue/yql_issue_manager.h>
 #include <yql/essentials/public/udf/udf_data_type.h>
@@ -2953,9 +2954,6 @@ const TTypeAnnotationNode* GetSeqItemType(const TTypeAnnotationNode* seq);
 const TTypeAnnotationNode& GetSeqItemType(const TTypeAnnotationNode& seq);
 
 const TTypeAnnotationNode& RemoveOptionality(const TTypeAnnotationNode& type);
-
-TMaybe<TIssue> NormalizeName(TPosition position, TString& name);
-TString NormalizeName(const TStringBuf& name);
 
 } // namespace NYql
