+++ yql/essentials/core/sql_types/ut/ya.make
@@ -1,6 +1,8 @@
 UNITTEST_FOR(yql/essentials/core/sql_types)
+
 SRCS(
     match_recognize_ut.cpp
+    normalize_name_ut.cpp
 )
 
 PEERDIR(
