+++ yql/essentials/core/ut/ya.make
@@ -4,7 +4,6 @@ SRCS(
     yql_column_order_ut.cpp
     yql_expr_constraint_ut.cpp
     yql_expr_optimize_ut.cpp
-    yql_expr_type_annotation_ut.cpp
     yql_library_compiler_ut.cpp
     yql_opt_utils_ut.cpp
     yql_udf_index_ut.cpp
